Push password encoding back into pkcs12_key_gen.
With PKCS8_encrypt_pbe and PKCS8_decrypt_pbe gone in
3e8b782c0cc0d9621f622cf80ab1a9bcf442fa17, we can restore the old
arrangement where the password encoding was handled in pkcs12_key_gen.
This simplifies the interface for the follow-up crypto/asn1 split.

Note this change is *not* a no-op for PKCS#12 files which use PBES2.
Before, we would perform the PKCS#12 password encoding for all parts of
PKCS#12 processing. The new behavior is we only perform it for the parts
that go through the PKCS#12 KDF. For such a file, it would only be the
MAC.

The specification, if anything, actually supports our old behavior. RFC
7292 B.1 asserts this transformation is an encoding thing and that
PKCS#5 simply failed to specify an encoding and that PKCS#12 uses the
bizarre NUL-terminated UTF-16 thing.

However, OpenSSL aligns with our new behavior. Today, we are unable to
process PBES2 PKCS#12 files that it produces and, with this change, we
are. I am unable to get any variant to interoperate with NSS or macOS,
though they appear to have code for it.

PKCS#12 is enough of a disaster that I suspect PBES2 PKCS#12 files
simply don't exist and are impossible to interoperate with. Absent a
coherent story here, let's just try to match OpenSSL and what's simpler.

TODO - Try a bit harder to figure out what's going on with at least one
other implementation and then give up...

BUG=54

Change-Id: I9a1bb4d5e168e6e76b82241e4634b1103e620b9b