Update patch set 1
Abandoned

OK, let's drop this idea for now, as there are too many prerequisites. In particular, there are a few little steps that could be taken but which would then eventually need to be thrown away, or there's a big step that's too much for me to do now.

Sorry to take your time on this.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon