Update patch set 2
Abandoned

> They believe that we shouldn't need anything beyond getauxval on ARMv8 chips (getauxval predates the first Android release with ARMv8 support) 

I'm not sure that's true, as I vaguely remember that there was an issue where the bits that `getauxval()` reports were not always being set due to an error of omission in Android, which is why the cpufeatures library does certain things the way it does.

OTOH, it might be the case that that bug was that ARMv8 features were not reported to 32-bit processes on a 64-bit OS, and that it might not matter assuming that we don't care too much about 32-bit processes on 64-bit Android. (Does anybody?)

In any case, thanks for answering the question.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon