Update patch set 2
Patch Set 2:

> Abandoned
> 
> > They believe that we shouldn't need anything beyond getauxval on ARMv8 chips (getauxval predates the first Android release with ARMv8 support) 
> 
> I'm not sure that's true, as I vaguely remember that there was an issue where the bits that `getauxval()` reports were not always being set due to an error of omission in Android, which is why the cpufeatures library does certain things the way it does.
> 
> OTOH, it might be the case that that bug was that ARMv8 features were not reported to 32-bit processes on a 64-bit OS, and that it might not matter assuming that we don't care too much about 32-bit processes on 64-bit Android. (Does anybody?)
> 
> In any case, thanks for answering the question.

I believe Chrome is currently a 32-bit process on 64-bit Android. :-)

Are you thinking a2d4c0c426caf2b212dea39d7bc05868d780eb34? That's actually not in cpufeatures (unless they've since fixed it), but it wouldn't be /proc/self/auxv. The problem is the AT_HWCAP2 auxv entry is missing altogether, so it's gone however you query auxv.

But, yes, sorry, I missummarized the thread. We shouldn't need anything beyond getauxval on ARMv8 chips to read the auxv. Separately, there's a bug in some configurations where auxv is missing an entry, so we have a non-auxv fallback for it.

Patch-set: 2