Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Maybe I'm missing something, but it seems to me that asynchronous ticket decryption is only useful if you also have asynchronous encryption. Is this wrong?

One could imagine perhaps using some kind of asymmetric scheme to avoid a round-trip on both legs of the handshake (or worse in TLS 1.3 because there are multiple tickets). We don't have a need for asynchronous encryption, so we didn't add it, to avoid introducing more complexity.

Note that supporting it is not as simple as this CL. Notably, this actually just causes a server infinite loop in TLS 1.3, and has some weird behaviors around the ticket_age_add business on retry.

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>