Randomise the high-bit of X25519 public values.
Compliant implementations of X25519 should mask the high bit of the
public key because the field is GF(2^255-19) but it's encoded in 256
bits, thus leaving a bit spare.

In order to ensure this happens, this change duplicates another bit from
the public key into the high-bit. This will make it appear to be
“random”.

Change-Id: Id7d79fdf5f106cc6e88ced42f85230b60e2ef393