Update patch set 1
Patch Set 1:

Somebody can build a statistical model to identify whether or not the high bit is likely to contain real information by comparing the frequency in which these two bits match. If these two bits always match then the high bit isn't storing information and if these two bits don't always match then either the high bit is random or it is encoding some information.

Maybe it would be better for the public_from_private function to set the high bit to one of the actually-random masked-off bits of the private key? That way it would be actually random, but without requiring any additional calls to the RNG.

Patch-set: 1
CC: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>