Update patch set 1
Patch Set 1:

> Somebody can build a statistical model to identify whether or not
 > the high bit is likely to contain real information

This is purely intended to catch implementations that are not correctly masking off the high bit on the basis that anything that isn't tested will be broken. Implementations that are /trying/ to be broken will manage it :) (I.e. this patch could just do public[31] |= 0x80. Perhaps it should?)

Patch-set: 1
Reviewer: Gerrit User 5070 <5070@66b711a4-aa9b-3dac-9062-916540830212>