Update patch set 1
Patch Set 1:

Note that you wrote another patch before for the *private* key that does always just set the high bit.

My understanding is that some people do intend to store a bit of information here so it does seem better to make it actually random so that we can argue about "indistinguishable from random" for those things. It so happens that we throw away 4 bits (IIRC) so one could be used to randomize the high bit of the private key, one could be used to randomize the high bit of the public key, and still have a couple of bits left over.

if people aren't going to actually store a value in this high bit then...not sure why it wasn't specified to be fixed to zero or one.

Patch-set: 1