Update patch set 1
Patch Set 1: Code-Review+1

lgtm, though it is odd that the public and private anti-canonicalizations are different. This one is random and the private one always sets to one. What's the reason for one vs. the other?

(My gut feeling is deterministic failures are better to avoid auto-reload troubles, but 50% may be a high enough probability that we're okay. And always setting it to one does risk people requiring it to be one, but that seems a little odd?)

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+1