Support Ed25519 in TLS.
This only works at TLS 1.2 and above as, before TLS 1.2, there is no way
to advertise support for Ed25519 or negotiate the correct signature
algorithm. Add tests for this accordingly.

For now, this is disabled by default on the verifying side but may be
enabled per SSL_CTX. Chromium, in particular, currently uses an external
verifier and will need a deployment plan. (On the signing side, we'll
happily enable it if given an Ed25519 certificate, so there's no need
for an additional toggle.)

BUG=187

Change-Id: Id25b0a677dcbe205ddd26d8dbba11c04bb520756