Update patch set 1
Patch Set 1:

(1 comment)

Please separate out server.cc from the others. We usually update the corpora in a separate change, just so they can be ignored more easily.

Also justify in the commit message why it's okay to cut corners on the teaching the fuzzer to toggle the settings per input. (I would prefer we just did it properly. We've been missing a lot of fuzzer coverage due to this for a while.)

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>