Update patch set 3
Patch Set 3:

I got this to build just fine. Clearly it's preferable to do this with the built-in iOS support in cmake. Can you elaborate on the problems with the problems the fat binary causes? Also, this version does not contain the armcap defines to enable HW crypto.

Patch-set: 3
Reviewer: Gerrit User 7805 <7805@66b711a4-aa9b-3dac-9062-916540830212>