Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I got this to build just fine. Clearly it's preferable to do this with the built-in iOS support in cmake. Can you elaborate on the problems with the problems the fat binary causes?

It seems the way Apple wants you to make a fat binary is that every single build operation gets passed multiple -arch flags. So all the way down to each individual .o file, everything is fat. Notably, this means the CMake file does not get to have a single notion of the target architecture. This is the same as the GYP limitations we were talking about.

This is not consistent with how our build is set up. Since the CMake build is only relevant for development, this seems a workable limitation for now. This would not affect the GN build.

> Also, this version does not contain the armcap defines to enable HW crypto.

Yours doesn't either, does it? I figured we'd sort that out after getting the tests running. I'm just trying to get things building for now.

Patch-set: 3
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>