Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > This is not consistent with how our build is set up. Since the
>  > CMake build is only relevant for development, this seems a workable
>  > limitation for now. This would not affect the GN build.
>  
> Ah, thanks for the clarification. We have this limitation with GN as well, and work around it with a script that builds all arches and lipos them together at the end.
> 
>  > Yours doesn't either, does it? I figured we'd sort that out after
>  > getting the tests running. I'm just trying to get things building
>  > for now.
> 
> Right, I left that out of the CL as it doesn't belong. 
> 
> I can work on getting at least the gtests running on device if you want to land this, don't know if you need support for that before landing this CL, given that we've been using boringSSL on iOS in webrtc and in chrome for a while now. IMO tests are critical for the asm stuff, though maybe not for this.

Nah, I'm happy landing this CL as-is. This is already including some untested stuff (the assembly you get without any bits flipped), but it's sort of hard to test things without it building. :-)

If it ends up being too much of a headache to get GTest and iOS to like each other standalone, we can switch to doing it on the Chromium roll. We would need to rejigger gtest_main.cc a bit (I'm thinking the embedder gets to swap out main but still is on the hook for calling some setup function), but standalone means we're less likely to have problems if, say, BoringSSL with an iOS bug rolls into the internal repository before it rolls into Chromium.

Patch-set: 3
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>