Be less clever about .rel.ro avoidance.
This restores the original version of delocate.go, with the subsequent
bugfixes patched in. With this, the FIPS module builds with GCC and
Clang, with and without optimizations.

Playing games with sections was a little overly clever and relied on the
compiler not performing a number of optimizations. Clang blew threw all
of those assumptions.

TODO digests.c is really wordy. Hide some of this behind a macro since
we'll probably have more method tables to come.

Change-Id: Ib4da468a5925998457994f9e392cf0c04573fe91