Update patch set 3
Patch Set 3:

FYI: A contributor to *ring* did some measurements and found that the optimization that was removed in this patch made a non-trivial difference in the performance of RSA blinding. Accordingly, in *ring* we decided to split `BN_rand_range_ex` into two functions: One that incorporates the the optimization, which is used only for RSA blinding, and one that doesn't, which is used for ECC (and eventually RSA) keygen.

Patch-set: 3
CC: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>