Update patch set 3
Patch Set 3:

> Okay, here, this is much simpler:

Indeed. I thought that I had tried that and rejected it, which was why it was the function hack. But I think I was testing in a little C file on the side and wasn't building with -fvisibility=hidden, which is why that didn't do the right thing for me. I've replaced this change with that one.

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>