Replace OPENSSL_ia32cap_addr with OPENSSL_ia32cap_addr_delta.
OPENSSL_ia32cap_addr avoids any relocations within the module, at the
cost of a runtime TEXTREL, which causes problems in some cases.
(Notably, if someone links us into a binary which uses the GCC "ifunc"
attribute, the loader crashes.)

This replaces assembly uses of OPENSSL_ia32cap_addr with a more
complicated scheme when building in FIPS mode. The
OPENSSL_ia32cap_addr_delta symbol (which is reachable relocation-free
from the module) stores the difference between OPENSSL_ia32cap_P and its
own address. Thus the module can find OPENSSL_ia32cap_P with a LEAQ and
ADDQ combination.

We cannot easily emit OPENSSL_ia32cap_addr_delta cross-platform (Windows
and macOS toolchains seem to dislike the 64-bit "PC-relative"
relocation). It is also not a 100% safe transformation as ADDQ clobbers
flags. Thus, it is implemented as a perlasm pseudo-instruction that
expands to a BORINGSSL_FIPS #ifdef. When patching the assembly, we
should check that the flags register may be safely clobbered.

Change-Id: I75497f02b2794560ad2aee24f11426ecb3ea50dd