Update patch set 13
Patch Set 13:

> This has a couple bad interactions with generate_build_files:
 > 
 > * cavp_test_util.cc winds up in crypto_sources, which (as a C++
 > file) I think it should not.
 > 
 > * There's a diagnostic message:
 > 
 > Test sources and configured tests do not match
 > These test sources are not configured: set(['cavp_aes_test',
 > 'cavp_aes_gcm_test'])

Yep, I plan on moving it to the top-level to avoid this. However I don't want to stomp on the work that people are currently doing so hoping to do that quickly over the weekend.

Have you hit this already?

Patch-set: 13
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>