Update patch set 1
Patch Set 1:

I think this script is getting too complicated and it's high time we redid it in Go. Or, better yet, we should be productionizing it (which will involve rewriting it anyway because the bots aren't going to run this shell script).

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>