Support more complex offset / symbol section interleavings.
LLVM likes to emit offsets of the form foo@toc@ha+16, which we didn't
support. Generalize parseMemRef to handle this case and avoid some of
the repeated offset special-cases. Offsets are now always folded into
the SymbolRef.

This still does not quite implement a fully general GAS-compatible
parser as GAS's parser is insane. GAS in x86_64 will happily accept
things like:

  1@GOTPCREL+foo
  blah1@GOTPCREL-blah2+blah3-blah4+blah5 # GOTPCREL modifies blah5, rest
                                         # of expression is an offset.

GAS actually textually pulls @GOTPCREL out of the input line partway
through parsing the expression and parses it afterwards! Then its normal
expression goes and maintains a running expression of a specific type
and, at each term, attempts to merge it into what it currently has. So
adding and subtracting symbols is not commutative (signs must
alternate) and the last symbol wins.

However its PPC64 parser does not allow this sort of transformation and,
instead, terminates each expression after @toc@ha and the like, except
that it special-cases foo@toc@ha+16: if we can parse one more expression
after @toc@ha AND it is a constant expression, then it is added into the
running offset. Otherwise it leaves that data unconsumed.

This is all ridiculous, so just generalize our parser slightly to cover
foo@toc@ha+16 and see how far we get from there.

Change-Id: I65970791fc10fb2638fd7be8cc841900eb997c9c